refactor litttle things
